refactor : move course classes to package
